perfect

+++ project/src/Types.cpp
@@ -14,7 +14,7 @@ std::ostream& operator<<(std::ostream& out,const Integer& value)
    return out;
 }
 //---------------------------------------------------------------------------
-Integer Integer::castString(const char* str,uint32_t strLen)
+Integer Integer::castString(const char* str,uint64_t strLen)
    // Cast a string to an integer value
 {
    auto iter=str,limit=str+strLen;
@@ -113,7 +113,7 @@ std::ostream& operator<<(std::ostream& out,const Date& value)
    return out << buffer;
 }
 //---------------------------------------------------------------------------
-Date Date::castString(const char* str,uint32_t strLen)
+Date Date::castString(const char* str,uint64_t strLen)
    // Cast a string to a date
 {
    auto iter=str,limit=str+strLen;
@@ -160,7 +160,7 @@ Date Date::castString(const char* str,uint32_t strLen)
    return d;
 }
 //---------------------------------------------------------------------------
-Timestamp Timestamp::castString(const char* str,uint32_t strLen)
+Timestamp Timestamp::castString(const char* str,uint64_t strLen)
    // Cast a string to a timestamp value
 {
    if ((strLen==4)&&(strncmp(str,"NULL",4)==0))
